requirements.txt
mammoth
requests
python-dotenv
python-docx
